added load_model, created model.py with all of the functions
