+++ config.yaml
@@ -2,16 +2,19 @@ setup:
   path: models/bee-world/
   gym_name: BeeWorld
   algorithm: TD3
-  alias: round3
-  continue_training: True
-  old_alias: round2
+  alias: round2
+
+  continue_training: False
+  old_alias: round1
 
 train: 
   timesteps: 2000
   iter_max: 10
   policy_kwargs: { net_arch: [100 ,100], activation_fn: ReLU }
   learning_rate: 0.001
+  video: True
 
 test:
-  prediction_step: 1000
+  prediction_steps: 1000
+  video: True
